[DEV] major update
