+++ palma/components/__init__.py
@@ -16,7 +16,7 @@ from palma.components.data_profiler import ProfilerYData
 from palma.components.dashboard import ExplainerDashboard
 from palma.components.performance import RegressionAnalysis
 from palma.components.performance import ScoringAnalysis
-from palma.components.performance import ShapAnalysis
+from palma.components.shap import ShapAnalysis
 from palma.components.data_checker import DeepCheck, Leakage
 
 __all__ = [
