+++ pyproject.toml
@@ -10,29 +10,11 @@ authors = [{ name = 'Vincent Laurent', email = 'vlaurent@eurobios.com' }]
 readme = "README.md"
 
 requires-python = '>=3.9'
-dependencies = [
-    'seaborn>0.12.0',
-    "tabulate>0.8.10",
-    "frozendict >2.3.4",
-    "flaml[automl]>=2",
-    "matplotlib>=3.4",
-    "numpy >= 1",
-    "scikit-learn >= 1, <1.4 ",
-    "pandas >= 1",
-    "shap",
-    "llvmlite >= 0.39",
-    "xgboost >2",
-    "pyaml>12",
-    "llvmlite",
-    "numba",
-]
-
+dependencies = [] # Dependencies will be loaded from requirements.txt
 
 [project.optional-dependencies]
-components = ["deepchecks", "ydata-profiling", "plot-metric",
-              "mlflow", "boto3", "explainerdashboard"]
-test = ["pylint", "coverage", "pytest", "pytest-cov", "pytest-cov",
-        "jupyter","notebook", "uci_dataset"]
+components = {file = ["requirements_components.txt"]}
+test = {file = ["tests/requirements.txt"]}
 doc = ["sphinx", "sphinx-book-theme", "myst-parser", "sphinx-copybutton",
        "sphinx-favicon", "nbsphinx",
        "sphinx-autoapi", "sphinx-version-warning", "sphinx-prompt",
