+++ requirements.txt
@@ -1,0 +1,14 @@
+seaborn>0.12.0
+tabulate>0.8.10
+frozendict>2.3.4
+flaml[automl]>=2
+matplotlib>=3.4
+numpy>=1
+scikit-learn>1.5
+pandas>=1
+shap
+llvmlite>=0.39
+xgboost>2
+pyaml>12
+llvmlite
+numba
