+++ requirements_components.txt
@@ -1,0 +1,7 @@
+deepchecks
+ydata-profiling
+plot-metric
+mlflow
+boto3
+explainerdashboard
+anywidget
